+++ src/Context.php
@@ -13,6 +13,7 @@ use DecodeLabs\Atlas;
 use DecodeLabs\Atlas\File;
 use DecodeLabs\Chronos;
 use DecodeLabs\Chronos\Blueprint\Factory as BlueprintFactory;
+use DecodeLabs\Chronos\Blueprint\Validation\Result as ValidationResult;
 use DecodeLabs\Veneer;
 
 class Context
@@ -29,6 +30,40 @@ class Context
 
         $factory = new BlueprintFactory();
         return $factory->load($file);
+    }
+
+    /**
+     * Load blueprint string
+     */
+    public function loadBlueprintString(
+        string $json
+    ): Blueprint {
+        $factory = new BlueprintFactory();
+        return $factory->loadString($json);
+    }
+
+    /**
+     * Validate blueprint
+     */
+    public function validateBlueprint(
+        string|File $file
+    ): ValidationResult {
+        if (is_string($file)) {
+            $file = Atlas::file($file);
+        }
+
+        $factory = new BlueprintFactory();
+        return $factory->validate($file);
+    }
+
+    /**
+     * Validate blueprint string
+     */
+    public function validateBlueprintString(
+        string $json
+    ): ValidationResult {
+        $factory = new BlueprintFactory();
+        return $factory->validateString($json);
     }
 }