+++ stubs/DecodeLabs/Chronos.php
@@ -10,6 +10,7 @@ use DecodeLabs\Veneer\ProxyTrait as ProxyTrait;
 use DecodeLabs\Chronos\Context as Inst;
 use DecodeLabs\Atlas\File as Ref0;
 use DecodeLabs\Chronos\Blueprint as Ref1;
+use DecodeLabs\Chronos\Blueprint\Validation\Result as Ref2;
 
 class Chronos implements Proxy
 {
@@ -22,5 +23,14 @@ class Chronos implements Proxy
 
     public static function loadBlueprint(Ref0|string $file): Ref1 {
         return static::$instance->loadBlueprint(...func_get_args());
+    }
+    public static function loadBlueprintString(string $json): Ref1 {
+        return static::$instance->loadBlueprintString(...func_get_args());
+    }
+    public static function validateBlueprint(Ref0|string $file): Ref2 {
+        return static::$instance->validateBlueprint(...func_get_args());
+    }
+    public static function validateBlueprintString(string $json): Ref2 {
+        return static::$instance->validateBlueprintString(...func_get_args());
     }
 };